+++ backend/internal/api/user.go
@@ -1,0 +1,30 @@
+package api
+
+import (
+	"github.com/gorilla/mux"
+	"github.com/sirupsen/logrus"
+	"net/http"
+	"shift/internal/entity"
+)
+
+func (s *APIServer) initUserRoutes(routes *mux.Route) {
+	routes.HandlerFunc(makeHTTPHandleFunc(s.handleCreateUser)).
+		PathPrefix("/users").
+		Methods(http.MethodPost)
+}
+
+func (s *APIServer) handleCreateUser(w http.ResponseWriter, r *http.Request) error {
+	logrus.Debugln("Create user handler running")
+
+	req := new(entity.CreateUserRequest)
+	if err := req.FromFormData(r); err != nil {
+		return BadRequestError{Message: err.Error()}
+	}
+
+	res, err := s.userService.CreateUser(req)
+	if err != nil {
+		return InternalServerError{Message: err.Error()}
+	}
+
+	return WriteJSONResponse(w, http.StatusOK, res)
+}
